feat: implement signal subsystem, process group management, and debugging tools
Summary of changes:
1. Signal Subsystem:
   - Implemented signal generation and delivery for: SIGCHLD, SIGALRM, SIGTERM,
     SIGKILL, SIGINT, SIGPIPE, SIGSEGV, and SIGILL.
   - Added support for user-defined signal handlers with sigreturn protocol.
   - Added signal-related syscalls: sys_signal, sys_sigaction, sys_kill,
     sys_sigpending, sys_sigprocmask, sys_pause, and sys_sigreturn.

2. Process & Memory Management:
   - Implemented process group support: sys_setpgid, sys_getpgid.
   - Added sys_waitpid for robust process lifecycle management.
   - Implemented sys_alarm for timed signal delivery.
   - Added initial sys_brk stub to task_struct for future memory refactoring.

3. TTY & Shell Enhancements:
   - Implemented foreground TTY management, enabling Ctrl+C (SIGINT) to
     terminate active user processes.
   - Refactored Shell logic to improve pipeline support and I/O control (sys_ioctl).

4. Debugging Infrastructure:
   - Enhanced PANIC macro to support stack backtracing.
   - Added 'addr2line.py' script to resolve kernel addresses into function names
     using the kernel.map file.

5. Task Structure:
   - Expanded task_struct with attributes for signal bitmasks, sigactions,
     and heap management.

+++ lib/stdint.h
@@ -10,5 +10,5 @@ typedef unsigned char uint8_t;
 typedef unsigned short int uint16_t;
 typedef unsigned int uint32_t;
 typedef unsigned long long int uint64_t;
-typedef int16_t pid_t;
+typedef int32_t pid_t;
 #endif
